Problem + Data
This excel sheet contains the problem and the data needed to complete the simulation. The Sales and Lead Times sheets are the data given by the organization. The DayleDemand sheet was generated from the Sales sheet, this data is used in the solution to generate random Dayle demand. The LeadTimeData sheet was generated from the Lead Times data, this data is used in the solution to generate random lead times for orders.